+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 - Fix some deadlock situations when the Application is unresponsive or stopped in debugger
+- Make connection timeout configurable; use balloon notification instead of exception in case of timeout
 
 ## [2.0.2]
 - (hopefully) fix agent loading bug on Windows